Tab on user profile to show starred repos
