+++ conf/locale/locale_en-US.ini
@@ -36,6 +36,7 @@ admin_panel = Admin Panel
 account_settings = Account Settings
 settings = Settings
 your_profile = Your Profile
+your_starred = Your starred
 your_settings = Your Settings
 
 activities = Activities
