+++ templates/base/head.tmpl
@@ -109,6 +109,10 @@
 													<i class="octicon octicon-person"></i>
 													{{.i18n.Tr "your_profile"}}<!-- Your profile -->
 												</a>
+												<a class="item" href="{{AppSubUrl}}/{{.SignedUser.Name}}?tab=stars">
+													<i class="octicon octicon-star"></i>
+													{{.i18n.Tr "your_starred"}}
+												</a>
 												<a class="{{if .PageIsUserSettings}}active{{end}} item" href="{{AppSubUrl}}/user/settings">
 													<i class="octicon octicon-settings"></i>
 													{{.i18n.Tr "your_settings"}}<!-- Your settings -->
